[css-grid] Ignore previously computed auto margins when relayout
This new NG layout engine implements a different approach for margin
calculation, giving the containing block the responsibility of its
children's margin computation.

However, in legacy layout this is children's responsibility; this is
the reason why the issue this patch tries to fix is only only
reproducible using layout-ng.

The grid code assumed this legacy behavior when computing the 'auto'
margins of the grid items. So, in order to fix the bug, this patch
just ignores any previously computed 'auto' margin when performing a
a relayout of a grid cntainer.

Bug: 994300
Change-Id: I44658876f7c2f2067202e45c6a552593bddfe6c9