[LayoutNG] Fixes positions/sizes of empty inline boxes
This patch sets to create a box fragment (not to cull) empty
inline boxes.

|NGInlineBoxState| already computes positions/sizes of empty
inline boxes if it were set not to be culled (e.g., by setting
background), but the emptiness was not included in the
condition to set |ShouldCreateBoxFragment()|.

According to crbug.com/997705, legacy works in some cases,
but not in any of simplified test cases nor in other issues.

Bug: 680075, 953479, 997705
Change-Id: I1726d1a573e34d88cc91b1950bfa8e327966833c