OOR-CORS: Make the CORS preflight have Accept: */*
The current spec does not require the CORS preflight request to
have an Accept header. But Firefox sends it, and Chrome sent it
until OOR-CORS is enabled.

https://github.com/whatwg/fetch/issues/922
Since there is no reason not to have the header, and no objections
at the spec proposal thread, we'll restore the Accept header to
the CORS preflight request.

Bug: 995740
Change-Id: I50854e65770df593c8bbcff7ee85404051a23f36